coments

+++ trunk/ex3-os1-2011/ex1b.c
@@ -3,23 +3,24 @@
 #include "shell.h"
 #include "ex1b.h"
 
+
 //=============================================================================
 int main()
 {	
 	setHandler();
-	setHendlerOptions();
-	cycle();
-	puts("Bye-Bye");		
-	return(EXIT_SUCCESS);
+	setHendlerOptions();			//	set handler
+	cycle();						//	enter to main cycle-function in shell
+	puts("Bye-Bye");				//	put bye bye
+	return(EXIT_SUCCESS);			//	exit
 }
 //================== Catch exit Handler =======================================
 void catch_chld(pid_t num)
 {
 
-	if(!WIFSTOPPED(status))
+	if(WIFEXITED(status))
 	{
-		wait4(-1,&status,WNOHANG,&u_rusage) ;
-		stoped_id = 0;
+		wait3(&status,WNOHANG,&u_rusage) ;
+		
 		long sys_time = u_rusage.ru_stime.tv_sec;
 		long sys_timeu =  u_rusage.ru_stime.tv_usec;
 		long usr_time = u_rusage.ru_utime.tv_sec;
@@ -28,6 +29,7 @@ void catch_chld(pid_t num)
 		printf("%ld.%06ld ,%ld.%06ld ,%d\
 		\n",sys_time,sys_timeu,usr_time,usr_timeu,exit_stat);
 	}
+
 }
 
 //=============================================================================
@@ -41,8 +43,8 @@ void catch_int(int num)
 //=============================================================================
 void catch_stop(int num)
 {
+	backgrdnd = 1;
 	setHandler();
-	//printf("\n");			//	print new line for neauty
 }
 
 //=============================================================================
@@ -62,12 +64,11 @@ void setHendlerOptions()
 //=============================================================================
 void cycle()
 {
-
-	char 	input[MAX_INPUT_LEN];
-	int 	size=0;	
-	char 	**vector_param = NULL;
-	int 	multi_task = 0;
-	pid_t 	child_pid;	
+	char 	input[MAX_INPUT_LEN];			//	variable for input
+	int 	size				=	0;		//	size of string array
+	char 	**vector_param 		= 	NULL;	//	string array
+	int 	multi_task 			= 	0;		//	have multi task process now
+	pid_t 	child_pid;						//	child process id
 	
 	while(getstring(input,MAX_INPUT_LEN))
 	{
@@ -76,13 +77,12 @@ void cycle()
 			break;
 		else if(!strcmp(input,"bg"))
 		{
-			if(stoped_id)
+			if(backgrdnd == 1 && stoped_id != 0)
 			{
+				backgrdnd = 0;
 				kill(stoped_id,SIGCONT);
-			
-				printf("Try %d\n",stoped_id);
-			
 				wait3(&status, WUNTRACED,&u_rusage);
+				setHandler();
 			}
 			
 			continue;
@@ -100,7 +100,10 @@ void cycle()
 		child_pid = fork();
 		
 		if(fork <0)
-			exit(EXIT_FAILURE);
+		{
+			perror("Can not fork()\n");	//	print error
+			exit(EXIT_FAILURE);			//	exit
+		}
 		else if(child_pid == 0)
 			exec(vector_param,size);	//	do execvp with vector param
 		else if(child_pid > 0)
@@ -109,13 +112,21 @@ void cycle()
 			free_arr(vector_param,size);
 			if(!multi_task)
 			{
-				stoped_id = child_pid;
+				if(backgrdnd == 0)
+				{
+					//backgrdnd = 0;
+					stoped_id = child_pid;
+					setHandler();
+				}
+				else
+					signal(SIGTSTP,SIG_IGN);
+					
 				wait3(&status, WUNTRACED,&u_rusage);
+				
 			}
 
 		}		
 	}	
 
 }
-
-
+//======================= END OF FILE =========================================
